chore: rename update-wallet-component to index
